'誰がログインしているかを判定する機能を追加'

+++ app/api/item/delete/[id]/route.js
@@ -3,11 +3,17 @@ import { ItemModel } from "@/app/utils/schemaModels";
 import { NextResponse } from "next/server";
 
 export async function DELETE(request, context) {
+  const reqBody = await request.json();
   try {
     await connectDB();
     const params = await context.params;
-    await ItemModel.deleteOne({ _id: params.id });
-    return NextResponse.json({ message: "アイテム削除成功" });
+    const singleItem = await ItemModel.findById(params.id);
+    if (singleItem.email === reqBody.email) {
+      await ItemModel.deleteOne({ _id: params.id });
+      return NextResponse.json({ message: "アイテム削除成功" });
+    } else {
+      return NextResponse.json({ message: "他の人が作成したアイテムです" });
+    }
   } catch (err) {
     return NextResponse.json({ message: "アイテム削除失敗" });
   }
